requirements.txt
Flask>=3.1.0
Flask-RESTful>=0.3.10
Flask-SQLAlchemy>=3.1.1
Flask-JWT-Extended>=4.6.0
Flask-Bcrypt>=1.0.1
Flask-CORS>=4.0.0
Flask-SocketIO>=5.4.0
SQLAlchemy>=2.0.36
python-dotenv>=1.0.0
reportlab>=4.0.0
requests>=2.32.0
